Add threshold support for levenshtein distance
I added threshold support for levenshtein distance. Although it increased the efficiency, I don't think it is the ideal solution. It looks the minumum value of the currently evaluated row in the levenshtein matrix as a stopping criteria since the next result will only add a cost to previous row's or column's result therefore if that minumum is bigger than the given threshold, which means the ultimate result is definetely going to be bigger than the threshold, the algorithm returns early with invalid result. Nevertheless, occasionally the minumum of all rows can be smaller than the threshold, but the result itself can be bigger. I couldn't come up a solution to that.