tr-levenshtein: Eliminate cost init. & encode overhead
Previous to this change, we have to initiliaze the costs everytime we call the c_levenshtein function. However c_levenshtein function is no-gil function. This means that we can not use a python object (such as a numpy array) on this function directly. Therefore, everytime we call this function, we would have to convert the numpy array to memoryview object. For small-scale project, this overhead is ignorable. Nevertheless, if you are planning to use the function billion times with same weight, it is preferable to initiliaze the costs (converting to memoryview object) beforehand the function call. This phenomena was also present for encoding words for every call. I also eliminate that with this change. These changes gave over 40% gain in speed performance.